🔨 reduce code duplication
